requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
alembic
pydantic
python-dotenv
requests
pgvector
sentence-transformers
nltk

# Vector search / embeddings
numpy
scikit-learn

# OpenAI support
openai

# Google Generative AI (Gemini)
google-generativeai